requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.2
gunicorn
django-crispy-forms==2.3
django-crontab==0.7.1
sqlparse==0.5.1
djangorestframework=3.15.2
